refactor(de): Split out template logic
